requirements.txt
autopep8==1.5.4
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.16.3
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
Pillow==7.2.0
plotly==4.11.0
pycodestyle==2.6.0
requests==2.24.0
retrying==1.3.3
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
